Allow getBaseUrl() and getModuleName() to be overridden
We currently use rebar-appcache for serving manifests, along with a scheme that encodes the user's chosen 
locale in the url of the manifest. For example, /MyApp/en.appcache or /MyApp/ar.appcache

Encoding the url in the appcache is preferable to using cookies because it allows users to have multiple languages
cached offline, and to have different languages open in different tabs, which is often needed during trainings.